+++ lib/pkm_level2_converter/pkm.rb
@@ -6,88 +6,35 @@ require 'pry'
 # Main Class
 class PKM
   def self.pools
-    return {
-    ausgangskontextPool: {
-      pathOfPool: '//xmlns:rn-tm/xmlns:ausgangskontext-pool | //xmlns:dl-km/xmlns:ausgangskontext-pool',
-      typeOfPool: AusgangskontextPool
-    },
-    ausgangsschnittstellenPool: {
-      pathOfPool: '//xmlns:rn-tm/xmlns:ausgangsschnittstelle-pool | //xmlns:dl-km/xmlns:ausgangsschnittstelle-pool',
-      typeOfPool: AusgangsschnittstellenPool
+    {
+      ausgangskontext_pool: {
+        path_of_pool: '//xmlns:rn-tm/xmlns:ausgangskontext-pool | //xmlns:dl-km/xmlns:ausgangskontext-pool',
+        type_of_pool: AusgangskontextPool
+      },
+      ausgangsschnittstellen_pool: {
+        path_of_pool: '//xmlns:rn-tm/xmlns:ausgangsschnittstelle-pool | //xmlns:dl-km/xmlns:ausgangsschnittstelle-pool',
+        type_of_pool: AusgangsschnittstellenPool
+      }
+      # Sprachen
     }
-    # Sprachen
-  }
-
-  def self.check_xsd(xsd, xml)
-    puts 'Check XML-Schema...'
-    xsd = Nokogiri::XML::Schema(xsd)
-    error = xsd.validate(xml)
-    if error.empty?
-      puts 'XML-Schema is valid.'
-      true
-    else
-      puts 'XML-Schema is invalid.'
-      error.each do |e|
-        puts e.message
-      end
-      false
-    end
   end
 
-  def self.parsePool(poolSymbol)
-    pool = @@pools[poolSymbol]
-    pathOfPool = pool[:pathOfPool]
-    typeOfPool = pool[:typeOfPool]
-    Pool.parse_pool(@xml_doc, pathOfPool, typeOfPool)
+  def self.parse_pool(pool_symbol)
+    pool = @@pools[pool_symbol]
+    path_of_pool = pool[:path_of_pool]
+    type_of_pool = pool[:type_of_pool]
+    Pool.parse_pool(@xml_doc, path_of_pool, type_of_pool)
   end
 
-  def self.xml_is_valid_pkm(xml)
-    # xsd =  File.read("./ka/pkm/1/XML-Schema_PKM.xsd")
-    path = ''
-    case xml.root.name
-    when 'pv-km'
-      path = './ka/pkm/1/XML-Schema_PKM.xsd'
-    when 'dl-km'
-      path = './ka/pkm/1/XML-Schema_PKM.xsd'
-    when 'rntm'
-      path = './ka/pkm/2/XML-Schema_PKM.xsd'
-    end
-
-    file_path = File.join(File.dirname(__FILE__), path)
-
-    xsd = File.read(file_path)
-    PKM.check_xsd(xsd, xml)
+  def initialize(pkm_xml)
+    @xml_doc = pkm_xml
+    @ausgangskontext_pool = PKM.parse_pool(:ausgangskontext_pool)
+    @ausgangsschnittstellen_pool = PKM.parse_pool(:ausgangsschnittstellen_pool)
   end
 
-  def self.convert_xpath_l3_id(node, x_path)
-    map = node.at_xpath x_path
-    id = map.content.to_i
-    map.content = (id + 0x8000).to_s
-  end
+  attr_reader :ausgangskontext_pool, :ausgangsschnittstellen_pool, :xml_doc
 
-  def initialize(file_name)
-    @filename = file_name
-
-    if File.file?(@filename)
-      @xml_doc = File.open(@filename) { |f| Nokogiri::XML(f) }
-    else
-      puts "File ('#{@filename}') not found."
-    end
-    # Is it a valid PKM file?
-    return unless PKM.xml_is_valid_pkm(@xml_doc)
-
-    parseXmlFile(@xml_doc)
-  end
-
-  def parseXmlFile(_xml)
-    # @xmlDoc = xml
-    @ausgangskontextPool = PKM.parsePool(:ausgangskontextPool)
-    @ausgangsschnittstellenPool = PKM.parsePool(:ausgangsschnittstellenPool)
-  end
-
-  attr_reader :ausgangskontextPool, :ausgangsschnittstellenPool, :filename, :xml_doc
-
-  def convertIds
+  def convert_org_id_to_level2
     case @xml_doc.root.name
     when 'pv-km'
       convert_ids_pvkm
@@ -98,107 +45,16 @@ class PKM
     else
       puts("#{@xml_doc.root.name} wird nicht unterstützt.")
     end
+
+    # TODO: CR 374
   end
 
-  def convertFileName
-    pn = Pathname.new(@filename)
-    dir, base = File.split(pn)
-    f = base.split('_')
-    org_id_l3 = f[1].to_i
-    org_id_l2 = org_id_l3 + 0x8000
-    f[1] = org_id_l2.to_s
-    Pathname.new(dir).join(f.join('_'))
-  end
-
-  def save_file
-    file = @filename
-    return unless PKM.xml_is_valid_pkm(@xml_doc)
-
-    # SAVE FILE with NEW name
-    new_file_name = convertFileName
-    output = File.open(new_file_name, 'w')
-    output << file.to_xml(indent_text: '', indent: 0).gsub(/>\n/, '>')
-    output.close
-    puts "File saved as: #{new_file_name}"
-  end
-
-  def convert_ids_dlkm
-    # Organisations-ID des DL:
-    # dl-km/organisation/id
-    convert_xpath_l3_id(@xml_doc, 'xmlns:dl-km/xmlns:organisation/xmlns:id')
-    puts 'Converted following IDs:'
-    puts "/dl-km/organisation/id = #{@xml_doc.xpath('/xmlns:dl-km/xmlns:organisation/xmlns:id').text}"
-
-    @xml_doc.xpath('/xmlns:dl-km/xmlns:kontrollmodul-pool/xmlns:item').each do |node|
-      # Organisations-ID des PV:
-      # dl-km/kontrollmodul-pool/item/moduldaten/organisation/id
-      convert_xpath_l3_id(node, 'xmlns:moduldaten/xmlns:organisation/xmlns:id')
-      puts "# dl-km/kontrollmodul-pool/item/moduldaten/organisation/id = #{node.xpath('xmlns:moduldaten/xmlns:organisation/xmlns:id').text}"
-      # Für PVKM zulässige Organisations-IDs der DL:
-      # dl-km/kontrollmodul-pool/item/moduldaten/organisation-pool/item/id
-      node.xpath('xmlns:moduldaten/xmlns:organisation-pool/xmlns:item').each do |child|
-        convert_xpath_l3_id(child, 'xmlns:id')
-        puts "## dl-km/kontrollmodul-pool/item/moduldaten/organisation-pool/item/id = #{child.xpath('xmlns:id').text}"
-      end
-    end
-    # Für Anzeige von KVP als Klartext:
-    # dl-km/kontrollmodul-pool/item/moduldaten/nummerninterpretation-pool/item[nr=2]/nummerntext-pool/item/nr
-  end
-
-  def convert_ids_pvkm
-    # Organisations-ID des DL:
-    # dl-km/organisation/id
-    # convert_xpath_l3_id(@xml_doc, 'xmlns:dl-km/xmlns:organisation/xmlns:id')
-    # puts "Converted following IDs:"
-    # puts "/dl-km/organisation/id = " + @xml_doc.xpath('/xmlns:dl-km/xmlns:organisation/xmlns:id').text
-
-    # @xml_doc.xpath('/xmlns:dl-km/xmlns:kontrollmodul-pool/xmlns:item').each do |node|
-    node = @xml_doc.xpath('/xmlns:pv-km')
-    # Organisations-ID des PV:
-    # dl-km/kontrollmodul-pool/item/moduldaten/organisation/id
-    # convert_xpath_l3_id(node, '/xmlns:pv-km/xmlns:organisation/xmlns:id')
-    convert_xpath_l3_id(node, 'xmlns:organisation/xmlns:id')
-    puts "# pv-km/organisation/id = #{node.xpath('xmlns:organisation/xmlns:id').text}"
-    # Für PVKM zulässige Organisations-IDs der DL:
-    # dl-km/kontrollmodul-pool/item/moduldaten/organisation-pool/item/id
-    node.xpath('xmlns:organisation-pool/xmlns:item').each do |child|
-      convert_xpath_l3_id(child, 'xmlns:id')
-      puts "## pv-km/organisation-pool/item/id = #{child.xpath('xmlns:id').text}"
-    end
-    # end
-    # Für Anzeige von KVP als Klartext:
-    # dl-km/kontrollmodul-pool/item/moduldaten/nummerninterpretation-pool/item[nr=2]/nummerntext-pool/item/nr
-  end
-
-  def convert_ids_rntm
-    # Organisations-ID des RN:
-    # rntm/herausgeber/nr
-    convert_xpath_l3_id(@xml_doc, 'xmlns:rntm/xmlns:herausgeber/xmlns:nr')
-    puts 'Converted following IDs:'
-    puts "rntm/herausgeber/nr = #{@xml_doc.xpath('xmlns:rntm/xmlns:herausgeber/xmlns:nr').text}"
-
-    @xml_doc.xpath('/xmlns:rntm/xmlns:tarifmodul-pool/xmlns:item').each do |node|
-      # Organisations-ID des PV:
-      # dl-km/kontrollmodul-pool/item/moduldaten/organisation/id
-
-      convert_xpath_l3_id(node, 'xmlns:tarifmodul/xmlns:herausgeber/xmlns:nr')
-      puts "# /rntm/tarifmodul-pool/item/tarifmodul/herausgeber/nr/ = #{node.xpath('xmlns:tarifmodul/xmlns:herausgeber/xmlns:nr').text}"
-      # Für PVKM zulässige Organisations-IDs der DL:
-      # dl-km/kontrollmodul-pool/item/moduldaten/organisation-pool/item/id
-      node.xpath('xmlns:tarifmodul/xmlns:organisation-pool/xmlns:item').each do |child|
-        convert_xpath_l3_id(child, 'xmlns:nr')
-        puts "## /rntm/tarifmodul-pool/item/tarifmodul/organisation-pool/item/nr = #{child.xpath('xmlns:nr').text}"
-      end
-    end
-    # Für Anzeige von KVP als Klartext:
-    # dl-km/kontrollmodul-pool/item/moduldaten/nummerninterpretation-pool/item[nr=2]/nummerntext-pool/item/nr
-  end
-
-  def convert_ids_AkDeutschlandticketPv
+  def convert_ids_cr374
     ## Ist die Ausgangsschnittstelle 3 bzw. 4 im Kontrollmodul vorhanden?
 
     # dl-km | rn-tm / ausgangsschnittstelle-pool / item / nr = 3 | 4
-    asst_pool_raw = @xml_doc.at_xpath('//xmlns:rn-tm/xmlns:ausgangsschnittstelle-pool | //xmlns:dl-km/xmlns:ausgangsschnittstelle-pool')
+    xpath = '//xmlns:rn-tm/xmlns:ausgangsschnittstelle-pool | //xmlns:dl-km/xmlns:ausgangsschnittstelle-pool'
+    asst_pool_raw = @xml_doc.at_xpath(xpath)
     asst_pool = AusgangsschnittstellenPool.new(asst_pool_raw)
 
     if !asst_pool_raw || asst_pool.ausgangsschnittstellen.empty?
@@ -206,11 +62,11 @@ class PKM
       return nil
     end
 
-    if !asst_pool.cr374?
+    if asst_pool.cr374?
+      puts 'CR374 ist umgesetzt. OrgIDs werden angepasst.'
+    else
       puts 'Das RN-TM / DL-KM hat keinerlei Aussgangsschnittstellen nach CR 374 definiert.'
       nil
-    else
-      puts 'CR374 ist umgesetzt. OrgIDs werden angepasst.'
     end
   end
 end
